Upgrade to Laravel 11

+++ app/Http/Controllers/RealtorListingController.php
@@ -3,24 +3,31 @@
 namespace App\Http\Controllers;
 
 use App\Models\Listing;
+use Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Inertia\Inertia;
+use Inertia\Response;
 
 class RealtorListingController extends Controller
 {
-    public function __construct()
-    {
-        $this->authorizeResource(Listing::class, 'listing');
-    }
+    // public function __construct()
+    // {
+    //     $this->authorizeResource(Listing::class, 'listing');
+    // }
 
     public function index(Request $request)
     {
+        Gate::authorize(
+            'viewAny',
+            Listing::class
+        );
         $filters = [
             'deleted' => $request->boolean('deleted'),
             ...$request->only(['by', 'order'])
         ];
 
-        return inertia(
+        return Inertia::render(
             'Realtor/Index',
             [
                 'filters' => $filters,
@@ -35,33 +42,38 @@ class RealtorListingController extends Controller
         );
     }
 
-    public function show(Listing $listing)
+    public function show(Listing $listing): Response
     {
-        return inertia(
+        Gate::authorize(
+            'view',
+            $listing
+        );
+        return Inertia::render(
             'Realtor/Show',
-            ['listing' => $listing->load('offers', 'offers.bidder')]
+            [
+                'listing' => $listing->load(
+                    'offers',
+                    'offers.bidder'
+                )
+            ]
         );
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
+    public function create(): Response
     {
-        // $this->authorize('create', Listing::class);
-        return inertia('Realtor/Create');
+        Gate::authorize(
+            'create',
+            Listing::class
+        );
+        return Inertia::render('Realtor/Create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
+        Gate::authorize(
+            'create',
+            Listing::class
+        );
         $request->user()->listings()->create(
             $request->validate([
                 'beds' => 'required|integer|min:0|max:20',
@@ -81,7 +93,11 @@ class RealtorListingController extends Controller
 
     public function edit(Listing $listing)
     {
-        return inertia(
+        Gate::authorize(
+            'update',
+            $listing
+        );
+        return Inertia::render(
             'Realtor/Edit',
             [
                 'listing' => $listing
@@ -91,6 +107,10 @@ class RealtorListingController extends Controller
 
     public function update(Request $request, Listing $listing)
     {
+        Gate::authorize(
+            'update',
+            $listing
+        );
         $listing->update(
             $request->validate([
                 'beds' => 'required|integer|min:0|max:20',
@@ -110,6 +130,10 @@ class RealtorListingController extends Controller
 
     public function destroy(Listing $listing)
     {
+        Gate::authorize(
+            'delete',
+            $listing
+        );
         $listing->deleteOrFail();
 
         return redirect()->back()
@@ -118,6 +142,10 @@ class RealtorListingController extends Controller
 
     public function restore(Listing $listing)
     {
+        Gate::authorize(
+            'restore',
+            $listing
+        );
         $listing->restore();
 
         return redirect()->back()->with('success', 'Listing was restored!');
